requirements.txt
PyYAML==6.0.1
simglucose==0.2.2
gym==0.9.4
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
matplotlib==3.5.0
matplotlib-inline==0.1.3
multiprocess==0.70.12.2
notebook==6.4.6
numpy==1.26.3
pandas==1.3.4
pyglet==1.5.21
python-dateutil==2.8.2
python-decouple==3.4
scipy==1.12.0
hydra-core==1.3.2
mlflow==2.14.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchvision==0.17.0+cu118
tornado==6.1

